1. waveForm 테스팅 준비 .

+++ soundlibrary/src/main/java/com/akakim/soundlibrary/gesture/OnDragGestureListener.kt
@@ -25,5 +25,5 @@ package com.akakim.soundlibrary.gesture
  */
 
 interface OnDragGestureListener{
-    fun onDrag(dx: Float, dy: Float)
+    fun onDrag(dx: Double, dy: Double)
 }
